+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master (v1.4.1)
 
+- chore(ping): improve code quality
 - build: create module for `network`
 - modules: update module files that uses `resolveIP()` function from `ping` module
 - core(network): move `resolveIP()` function to here