+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master (v1.4.1)
 
+- refactor(gpu): remove unnecessary `free()` functions
 - refactor(updates): remove unnecessary `EUID` allocating
 - refactor(updates): replace `fs.max_path_bytes` with `linux.MAX_PATH`
 - chore(ping): improve code quality